Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/registration-schedule-job.https-expected.txt


FAIL different scriptURL and updateViaCache promise_test: Unhandled rejection with value: object "TypeError: null is not an object (evaluating 'registration.installing.scriptURL')"
FAIL different type assert_equals: expected (string) "classic" but got (undefined) undefined
